Bumping libprofit to version 0.13
This version brings along the new profiles plus the centralization of the
common code among the radial profiles. This imposes a new naming in some of the
parameters:

 * A common "rscale" concept has been introduced (what in sersic was Re) which
   is derived from different parameters for different profiles
 * The already existing re_switch and re_max parameters have been renamed
   rscale_switch and rscale_max accordingly.

Only the re_max parameter was currently being exposed to R, so I took care of
performing the translation to avoid having to change the R interface again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>