Some major high level changes.
intervals now takes a simple list input with the limits as vectors. Does not have to be specified.

priors is a function that returns the log-likelihood itself. This means arbitrarily complex priors can be formed (conditional MVN etc). Does not have to be specified.

New constraints function that allows basically any manipulation of the modellist (so setting values to be multiples of others et al). Does not have to be specified.

Examples have been updates, and everything seems to work okay...