Removing extra backslash
There is nothing else to escape so there's no need for the final backslash.
Most make programs play happily with it (which is why we hadn't realize about
it at first), but the SunOS builds on CRAN were apparently failing due to this
(although the error messages we received gave no clue about the real underlying
reason).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>