Initial support for configurable compilation flags
The new configure script checks the system for OpenCL support and generates the
Makevars file accordingly. This has been tested in Linux only so far, so things
might go not so nicely in other platforms (of special interest is MacOSX where
we need to be aware of the OpenCL framework, etc).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>