profitSegImWatershed -> profitMakeSegim profitSegImExpand -> profitMakeSegimExpand
Some doc fixes to reflect the name change and a better example in profitMakeSegimExpand.

Changed lots of return() to return= to force the output to be silent (lots of the functions in profitUtility have plotting side effects, so you do not always want to keep the output and it is therefore better to make these silent).