Undoing explicit symbol listing for shared library
I didn't consider the rest of the functions defined in other source code files
that are still used via .Call(). These functions need to be put into the list
of exported symbols; otherwise they fail to load. I prefer not to rush things,
so for the time being I'll simply comment this code out.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>