Convolve when PSF is given
This is a small adjustment to the logic I had changed before: profile images
should be convolved any time a PSF is given. The convolver being present or not
doesn't alter this fact, it only allows users to use a specific convolver. The
Model will anyway create a convolver if one has not been given but is needed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>