Bringing in libprofit 1.6.0
This version includes OpenMP-accelerated CPU brute-force convolution,
plus a number of fixes and improvements, mostly related to OpenCL, that
we have found (and fixed temporarily) while running ProFit and the
benchmarks in pyprofit.

This commit overwrites completely the alternative brute-force
convolution code that Dan added some days ago. With the coming of the
OpenMP-accelerated code it was a bit difficult for me to integrate Dan's
code. We also observed some memory access problems when compiling that
code with clang (at least on a Mac, haven't tried on Linux). The
repository already contains the code in its history anyway, so we can go
back and incorporate it again if required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>