Using sub-optimal arithmetic expression evaluation
The "Writing R Extensions" manual is not clear in regards to the exact
syntax dialect that should be used in the configure script. On the one
hand they mention that "you can include an executable (Bourne) shell
script configure in your package", but the associated footnote mentions
that "The script should only assume a POSIX-compliant /bin/sh", with a
link pointing to [1], the Shell Command Language specification for POSIX
shells. In this last reference, the $((...)) syntax is explicitly
mentioned [2].

In reality CRAN submissions seem to need the Bourne shell syntax dialect
though (we were warned that the statements we are now "correcting" in
this commit did not use Bourne syntax). This forces us to change these
evaluations to use the expr command, which I don't think is a builtin
function of the Bourne shell, and therefore might or might not work in
different systems depending on whether the expr executable is found or
not.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html
[2] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_04

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>