Avoid creating lists with NAs
When users specify a single value for a property that otherwise should
be specified multiple times (because the "xcen" property in the profile
list is specified multiple times) we can asume they mean to use the same
value for all individual profiles. In the previous incarnation of the
code the original single-itemd lists were automatically expanded, which
resulted in NAs being added in places where proper values should have
been. This made the code non-deterministic, giving different results in
different machines.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>