Using up-/down-sampling from libprofit 1.7.0+
Sometimes the upsample/downsample methods are used with boolean mask
surfaces, in which case we still want to load them as a matrix of double
values (which is what the old code did), and therefore we now consider
that case.

I took the chance to change the parameter names of the two R functions
to be lowercased, and updated the documentation accordingly.

This change means that we no longer require the Rcpp-generated
profitDownsample/profitUpsample code. Actually, Rcpp is now not required
at all.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>