Reworking profitMakeModel for libprofit 1.7.2+
Since now libprofit takes care of dealing with the finesampling,
returnfine and returncrop arguments, we simply need to pass them down to
the call to C code instead of modifying all the different profiles'
parameters in R land to accommodate for them.

This takes away a lot of complexity from the profitMakeModel function,
which now only needs to take care of creating a well-structured list of
profiles for the given parameters.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>