Check for non-NULL OpenCL environments
There are a wide number of reasons why an OpenCL environment could fail
to be created, so we need to be aware of that and act accordingly by not
using the NULL result that comes out of the C world.

The documentation of profitOpenCLEnv has also been updated to reflect
this behavior, which was not explicitly stated before.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>