Enhancing OpenMP support checks
We were previously only checking that the OpenMP flag correctly compiled
files, but not that it correctly *linked* different objects into a
library or program. Even though is is extremely unlikely, this linking
step can still fail due to several reasons. We are therefore adding an
extra linking check step when testing for a correct OpenMP flag. If we
find ourselves in such a situation we print a warning message and
disable our OpenMP support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>