Fixing Windows compilation
The macro that signals we are building the library changed, but since we
always build in Linux platforms we missed this particular change. Sorry
about that!

Moreover, Windows uses a different file, so we need to update not the
configure script, but also the more static configuration we use in
Makevars.win

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>