Sorry, we are not supporting devel anymore
The main problem (so far) was always trying to get the fftw R package to
compile both in x86 and x64 architectures. We tried to compile against
two different installations of fftw: one using vcpkg and other via
manual download/unzip/mkdir from the official fftw site. In both cases
the problem was that the fftw R package uses only the FFTW_LIB env
variable to find an fftw installation, but AppVeyor's R builds its
packages for both architectures (and a single fftw installation cannot
be used for both).

There probably is a way to constrain the architecture on which the tests
run, but this was the development version of R anyway, so is not a big
deal.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>