Avoid POSIX expansions unsupported in Solaris
The Solaris /bin/sh shell does not implement some of the POSIX
replacements we were using for substring extraction, and therefore a
different route needs to be taken (since there seems to be no way to
specify that *we* don't want to support Solaris).

Instead, we are resorting to use sed, although there might more
efficient ways to do this. To make sure we run correctly in all
platforms we don't use extended regular expressions, so the matching
pattern looks quite daunting.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>