Bye El Capitan, hello High Sierra
The xcode7.3 images are not working anymore with the latest release of R
for MacOS because their tcltk package requires the latest XQuartz
2.7.11, the image comes with 2.7.9, and there are no plans to upgrade it
anytime soon (see travis-ci/travis-ci#8762). I tried to manually install
XQuartz 2.7.11 and this fixed the original error, but only to have R
hang up on me consistently on every build. I don't know exactly what's
going on, but I suspect that after installing XQuartz I need to get a
hold on the new DISPLAY variable for everything to work correctly. I
can't find any reference on how to do this automatically (all references
on the web simply talk about logging out and back in, or rebooting the
machine), so I'm giving up on this one.

To compensate for this I'm bringing up MacOS 10.13 High Sierra, which
should ensure some continuity on our MacOS support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>