Avoiding -Wunused-private-field warnings
There were a few OpenMP-related private fields that were causing
warnings to be issued by some compilers because they were not used when
the PROFIT_OPENMP macro is not defined. This patch attempts to remove
those warnings away. The fix upstream in libprofit is slightly different
because the code has changed a bit more, but the spirit of the fix is
the same.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>