Do not reset smart pointers
There should be no need to manually reset the shared pointers when
destructing their wrappers. Indeed, the wrapper destruction should
be enough as they will destruct their shared pointer members, which will
eventually delete the underlying object.

This might or might not have to do with the ERROR messages we are
getting in the MacOS CRAN builds, which we haven't been able to
reproduce anywhere else.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>