Moving libprofit headers to include/ dir
This is something I meant to do some time ago, but now it acquired some
urgency given that ProFit is failing to build on MacOS with clang 8.0.0.
I'm doing this locally, but I will also perform the same change in
libprofit upstream.

Due to this move, all references to the auto-generated config.h file
(which we placed under src/libprofit/profit, but now under
src/libprofit/include/profit) had to be updated to reflect the new
directory structure.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>