Correct offset calculation and application
There was a disconnect between how the offset was returned to the R
layer and how it was then used to calculate the X and Y labels
associated to the resulting image. This incongruence was probably caused
by me thinking that R is a 1-based language, which obviously has nothing
to do with how we label pixels in our images.

On top of this disconnect, the calculation of the sequence of numbers
generated one element too many, so I'm fixing that as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>