Bringing krn > img support from libprofit
This support is generic to all convolvers and could certainly be
refined, as it could be avoided more intelligently in the convolvers
that already do internal padding of their images. The use case should
happen rarely though, so it's not that urgent to perform well.

This code is living at the top of libprofit, which is currently sitting
in version 1.8.2. I thus rather manually bring this patch over instead
of the whole libprofit changes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>