Adding new adjust_calcregion flag
This flag tells the libprofit Model class whether the given calcregion
(Mask in libprofit parlance) should be internally adjusted or not to
make sure all the intended flux is captured as required. The flag is on
by default, and when users set it to false libprofit's Model still
checks that the given calcregion has the intended dimensions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>