OpenCL kernel header generation added to configure
Since libprofit 1.8 the way the OpenCL sources are embedded into the
libprofit library has changed; it previously used an awkward combination
of #include macros and include paths, while now the process is cleaner
and better defined in two stages: first, C++ header files are generated
from the .cl sources to define a global, constant variables containing
the OpenCL source code. And secondly, these headers are included in the
opencl.cpp file and the variables read to get the OpenCL sources.

OpenCL kernel header generation needs to only happen when OpenCL is
enabled of course, and these generated files need both to be ignored and
cleaned up.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>