Solving small compilation problems
These stem from the new libprofit version being pulled, which brings its
own profit::exception (and thus there was a cash of names) and its own
profit::detail namespace (which also clashed with the ::detail
namespace used in r_profit.cpp). In the case of the latter, the code
inside the ::detail namespace was actually not needed anymore since this
functionality is now provided by profit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>