Only include OpenCL kernel headers when needed
In libprofit we always generate the headers with the OpenCL kernel
source code, but in ProFit we are doing this only when OpenCL support is
found (which is probably the correct way of doing it). I thought the
code was prepared for this but it wasn't, as it actually failed to build
in AppVeyor (Windows).

This change should fix this problem, and is something I'll be putting in
upstream libprofit too.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>