Avoiding copies while caching OpenCL kernels
This was found while running libprofit's unit tests with valgrind, with
it complaining that the call to 'write' was using uninitialized bytes
(in the block of data to be written), which was a local copy of the
OpenCL kernel binary data. Not too big of an issue in ProFit (we don't
use caching really), but still: better safe than sorry.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>