Avoid warning due to missing "checkbashisms"
This is a script that newer R versions seem to require to be present in
order to check for bashisms in the configure script, but it doesn't come
(yet) installed by default in the machine images used by Travis CI's R
support. We are resorting to installing it manually until such support
comes in.

For details see https://travis-ci.community/t/new-warning-in-r-devel-about-the-checkbashisms-script/5551

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>