Updating MacOS versions in Travis build matrix
The old versions of Xcode meant we were still using MacOS 10.12
(Sierra), but homebrew doesn't provide pre-built binaries for all
packages (e.g., gcc, which is finally required by some of our direct
dependencies). Binaries packages require compilation, and compiling gcc
times out because it takes more than 10 minutes to finish, in which time
homebrew spits no output.

I'm upgrading the versions of Xcode we use so we use MacOS 10.13 and
10.14. This is probably more representative of the platforms we support
anyway, so it's not a bad change.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>