Avoid errors when looking up for FFTW3 libs
Newer compilers are more pedantic about function declarations having to
be present in order for a module to compile. This prevented FFTW from
being found on MacOS platforms.

At this point we know the fftw headers are present in the system, so we
are simply including them to get the function definition we need.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>