Apply the pixel size offset to xcen and ycen, too
This cannot be covered by offset[1] and offset[2] because when pixel sizes are different between images, there is no single additive offset that will get you from any given point in image1 to the same point in image2. 
When offset[4] is non-zero, this needs to be considered in specifying offset[1] and offset[2]. The way it is now is that the reference pixel scale for offset[1] and offset[2] is that of the image to which they are applied (rather than specifying all offsets in the pixel scale of the first image). 
We should probably warn users about this in the docs somewhere...