Doc update to reflect changes in treatment of pixel size offset
Warn users that the pixel scale offset is applied to xcen and ycen position, too and that this needs to be considered when specifying those. 
Note that in the corresponding code commit, I got the description mixed up a little: offset[1] and offset[2] should in fact be specified using the pixel scale of the first (reference) image, not the image to which they are applied.